move request count to UsageControl cmp
