+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/usage-bar/types.ts
@@ -9,20 +9,3 @@ export type UsageBarProps = {
 	 */
 	limitReached: boolean;
 };
-
-export type UsageControlProps = {
-	/**
-	 * The current usage, as a percentage represented by a number between 0 and 1.
-	 */
-	usage: number;
-
-	/**
-	 * True if the usage is over the limit.
-	 */
-	isOverLimit: boolean;
-
-	/**
-	 * True if the AI Assistant feature is supported.
-	 */
-	hasFeature: boolean;
-};
