+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/usage-panel/style.scss
@@ -1,7 +1,6 @@
 @import '@automattic/jetpack-base-styles/style';
 
 .jetpack-ai-usage-panel-control {
-    width: 100%;
 
     p.muted {
         color: var(--jp-gray-50 );
